# Profiler: Rewrite serializer to stream compressed output, reducing memory usage