--- a/profiling/src/collections/identifiable/mod.rs
+++ b/profiling/src/collections/identifiable/mod.rs
@@ -76,14 +76,6 @@ impl<T: Item> Dedup<T> for FxIndexSet<T> {
     }
 }
 
-pub fn to_pprof_vec<T: PprofItem>(collection: FxIndexSet<T>) -> Vec<T::PprofMessage> {
-    collection
-        .into_iter()
-        .enumerate()
-        .map(|(index, item)| item.to_pprof(<T as Item>::Id::from_offset(index)))
-        .collect()
-}
-
 pub fn into_pprof_iter<T: PprofItem>(
     collection: FxIndexSet<T>,
 ) -> impl Iterator<Item = T::PprofMessage> {