--- a/profiling/src/exporter/mod.rs
+++ b/profiling/src/exporter/mod.rs
@@ -1,15 +1,16 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
 
-use std::borrow::Cow;
 use std::future;
 use std::io::Cursor;
+use std::{borrow::Cow, io::Write};
 
 use bytes::Bytes;
 pub use chrono::{DateTime, Utc};
 pub use ddcommon::tag::Tag;
 pub use hyper::Uri;
 use hyper_multipart_rfc7578::client::multipart;
+use lz4_flex::frame::FrameEncoder;
 use mime;
 use serde_json::json;
 use tokio::runtime::Runtime;
@@ -53,6 +54,7 @@ pub struct ProfileExporter {
 pub struct File<'a> {
     pub name: &'a str,
     pub bytes: &'a [u8],
+    pub compress_before_exporting: bool,
 }
 
 #[derive(Debug)]
@@ -236,25 +238,26 @@ impl ProfileExporter {
         );
 
         for file in files {
-            //TODO : confirm that this is not needed since we now pre-encode
-            // // We tend to have good compression ratios for the pprof files,
-            // // especially with timeline enabled. Not all files compress this
-            // // well, but these are just initial Vec sizes, not a hard-bound.
-            // // Using 1/10 gives us a better start than starting at zero, while
-            // // not reserving too much for things that compress really well, and
-            // // power-of-two capacities are almost always the best performing.
-            // let capacity = (file.bytes.len() / 10).next_power_of_two();
-            // let buffer = Vec::with_capacity(capacity);
-            // let mut encoder = FrameEncoder::new(buffer);
-            // encoder.write_all(file.bytes)?;
-            // let encoded = encoder.finish()?;
+            let encoded = if file.compress_before_exporting {
+                // We tend to have good compression ratios for the pprof files,
+                // especially with timeline enabled. Not all files compress this
+                // well, but these are just initial Vec sizes, not a hard-bound.
+                // Using 1/10 gives us a better start than starting at zero, while
+                // not reserving too much for things that compress really well, and
+                // power-of-two capacities are almost always the best performing.
+                let capacity = (file.bytes.len() / 10).next_power_of_two();
+                let buffer = Vec::with_capacity(capacity);
+                let mut encoder = FrameEncoder::new(buffer);
+                encoder.write_all(file.bytes)?;
+                encoder.finish()?
+            } else {
+                file.bytes.to_vec()
+            };
             /* The Datadog RFC examples strip off the file extension, but the exact behavior isn't
              * specified. This does the simple thing of using the filename without modification for
              * the form name because intake does not care about these name of the form field for
              * these attachments.
              */
-            //TODO, take the file as owned
-            let encoded = file.bytes.to_vec();
             form.add_reader_file(file.name, Cursor::new(encoded), file.name)
         }