--- a/profiling/src/exporter/mod.rs
+++ b/profiling/src/exporter/mod.rs
@@ -1,9 +1,9 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
 
+use std::borrow::Cow;
 use std::future;
-use std::io::Cursor;
-use std::{borrow::Cow, io::Write};
+use std::io::{Cursor, Write};
 
 use bytes::Bytes;
 pub use chrono::{DateTime, Utc};