--- a/profiling/src/lib.rs
+++ b/profiling/src/lib.rs
@@ -4,3 +4,4 @@
 pub mod collections;
 pub mod exporter;
 pub mod profile;
+pub mod serializer;