--- a/profiling/src/profile/pprof/mod.rs
+++ b/profiling/src/profile/pprof/mod.rs
@@ -1,0 +1,2 @@
+mod proto;
+pub use proto::*;