# What does this PR do?

Reduces memory needed to serialize the profiler into a pprof. In particular, it enables serialization to occur in a streaming manner which consumes the profile, allowing memory usage to remain flat.

# Motivation

Currently, serialization takes significant memory (see below, ~60MB on a Ruby benchmark with timeline enabled).  The standard `prost` model of serializing to protobuf works on complete messages, and requires multiple copies of the data.  Instead, tweak the serialization to:

1. Consume the in-memory representation of the Profile as it is serialized
2. Stream the creation of the serialized pprof
3. stream lz4 encoding of the pprof (timeline pprofs compress extremely well, so keeping only the compressed version in memory saves memory).

Running @ivoanjo Ruby benchmark gives

<meta charset="utf-8"><b style="font-weight:normal;" id="docs-internal-guid-4f1ff639-7fff-dbc6-336e-b6d4bce873bf"><div dir="ltr" style="margin-left:0pt;" align="center">

Config | Timeline | Process Memory Rss before serialization (bytes) | Process Memory Rss after serialization (bytes) | Process Memory Rss after GC (bytes) | Maximum rss (from /bin/time -v, bytes)
-- | -- | -- | -- | -- | --
16 threads with depth 50 (libdatadog 4.0) | false | 70980k | 70980k | 70980k | 70980k
16 threads with depth 50(libdatadog 4.0) | true | 78636k | 134004k | 85196k | 153676k
16 threads with depth 50 (libdatadog 5.0) | false | 70900k | 71060k | 71060k | 71060k
16 threads with depth 50 (libdatadog 5.0) | true | 78808k | 76540k | 76540k | 79128k
4 threads with depth 400(libdatadog 5.0) | true | 76512k | 76136k | 76136k | 77472k

</div><br /></b>


# Additional Notes

* Builds on @ivoanjo work in https://github.com/DataDog/libdatadog/commit/3bb98dc5074e034ca8b26f50360e2291274c8652
* The key insight is that the **top level** protobuf message doesn't carry any frame information. So the emitted bytes are identical if we either (1) serialize the `Profile` top level message or (2) serialize a series of single field messages, each with the tag of their corresponding `Profile` field.  This only works for the top-level message, but luckily that's all we need.*

# How to test the change?
* Rerun the ruby benchmark, and see the memory savings
* I modified all tests to round-trip through the serializer, which ensures that both serialization and deserialization work. 

## For Reviewers
- [ ] If this PR touches code that signs or publishes builds or packages, or handles credentials of any kind, I've requested a review from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.