@ivoanjo 2023-09-13
Before reviewing the PR, I've spent some time updating the [Ruby timeline memory testing branch](https://github.com/datadog/dd-trace-rb/tree/ivoanjo/timeline-memory-experiments-v2) with the latest changes.

The test suite is 100% green with the new serializer, which I think is great news -- there's a lot of tests in the Ruby profiler test suite where we turn on the profiler, then serialize the result, then deserialize it and assert on it, and no issues there with the serializer.

I've also tested it in staging with our prof-fibonacci minimal test app, and both timeline, as well as regular profiles are in good shape. I also did some really quick memory measurements for prof-fibonacci:

* dd-trace-rb v1.14.0 (libdatadog 3) + timeline: maximum rss 151548 kb
* timeline testing branch (libdatadog 5) + timeline: maximum rss 99620 kb
* timeline testing branch (libdatadog 5) + no timeline: maximum rss 92812 kb