@ivoanjo 2023-09-13
P.s: Also probably not a bad idea to update the PR title to reflect that we also kinda rebuilt the serializer ;)